# apps/api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Edutech Platform API",
        "description": "API documentation for Edutech Platform - Bimbel & Tryout Platform",
        "contact": {
            "name": "Edutech Team",
            "email": "api@edutech.example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/api/v1/exams/{attempt}/log": {
            "post": {
                "tags": [
                    "CBT"
                ],
                "summary": "Log a proctoring event",
                "description": "Logs events like focus lost, tab switch, etc.",
                "operationId": "aa19898487a4fc0924bfa3060dd6ed45",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "focus_lost"
                                    },
                                    "meta": {
                                        "type": "object",
                                        "example": {
                                            "reason": "User switched tab"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Event logged"
                    },
                    "404": {
                        "description": "Attempt not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/exams/{attempt}/heartbeat": {
            "post": {
                "tags": [
                    "CBT"
                ],
                "summary": "Send a heartbeat",
                "description": "Updates the last activity timestamp of the attempt.",
                "operationId": "7989b3203088c9c2ff9a72c05aae51a0",
                "responses": {
                    "200": {
                        "description": "Heartbeat received"
                    },
                    "404": {
                        "description": "Attempt not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/learning/mentors": {
            "get": {
                "tags": [
                    "Learning"
                ],
                "summary": "List all mentors",
                "operationId": "c0c118a48e4636b550fa0202dd9202ad",
                "responses": {
                    "200": {
                        "description": "List of mentors"
                    }
                }
            },
            "post": {
                "tags": [
                    "Learning"
                ],
                "summary": "Create a mentor profile",
                "operationId": "4d3bc30f821c3167045484ed041e6718",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "specialization": {
                                        "type": "string"
                                    },
                                    "bio": {
                                        "type": "string"
                                    },
                                    "experience_years": {
                                        "type": "integer"
                                    },
                                    "social_links": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mentor created"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}